refactor, save checkpoint
