--- rooset-domain-query-reducers/npm-debug.log
@@ -1,48 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/home/daniel/.nvm/versions/node/v6.10.1/bin/node',
-1 verbose cli   '/home/daniel/.nvm/versions/node/v6.10.1/bin/npm',
-1 verbose cli   'run',
-1 verbose cli   'test' ]
-2 info using npm@3.10.10
-3 info using node@v6.10.1
-4 verbose run-script [ 'pretest', 'test', 'posttest' ]
-5 info lifecycle rooset-domain-query-reducers@0.1.0~pretest: rooset-domain-query-reducers@0.1.0
-6 silly lifecycle rooset-domain-query-reducers@0.1.0~pretest: no script for pretest, continuing
-7 info lifecycle rooset-domain-query-reducers@0.1.0~test: rooset-domain-query-reducers@0.1.0
-8 verbose lifecycle rooset-domain-query-reducers@0.1.0~test: unsafe-perm in lifecycle true
-9 verbose lifecycle rooset-domain-query-reducers@0.1.0~test: PATH: /home/daniel/.nvm/versions/node/v6.10.1/lib/node_modules/npm/bin/node-gyp-bin:/home/daniel/src/rooset/rooset-domain-query-reducers/node_modules/.bin:/home/daniel/.nvm/versions/node/v6.10.1/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin:/home/daniel/Android/Sdk/tools:/home/daniel/Android/Sdk/platform-tools:/home/daniel/src/my-react-native-utils:/home/daniel/src/my-utils:/home/daniel/.local/bin:/snap/bin:/home/daniel/Android/Sdk/tools:/home/daniel/Android/Sdk/platform-tools:/home/daniel/src/my-react-native-utils:/home/daniel/src/my-utils:/home/daniel/.local/bin
-10 verbose lifecycle rooset-domain-query-reducers@0.1.0~test: CWD: /home/daniel/src/rooset/rooset-domain-query-reducers
-11 silly lifecycle rooset-domain-query-reducers@0.1.0~test: Args: [ '-c', 'jest --runInBand' ]
-12 silly lifecycle rooset-domain-query-reducers@0.1.0~test: Returned: code: 1  signal: null
-13 info lifecycle rooset-domain-query-reducers@0.1.0~test: Failed to exec test script
-14 verbose stack Error: rooset-domain-query-reducers@0.1.0 test: `jest --runInBand`
-14 verbose stack Exit status 1
-14 verbose stack     at EventEmitter.<anonymous> (/home/daniel/.nvm/versions/node/v6.10.1/lib/node_modules/npm/lib/utils/lifecycle.js:255:16)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:191:7)
-14 verbose stack     at ChildProcess.<anonymous> (/home/daniel/.nvm/versions/node/v6.10.1/lib/node_modules/npm/lib/utils/spawn.js:40:14)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:191:7)
-14 verbose stack     at maybeClose (internal/child_process.js:886:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
-15 verbose pkgid rooset-domain-query-reducers@0.1.0
-16 verbose cwd /home/daniel/src/rooset/rooset-domain-query-reducers
-17 error Linux 4.4.0-66-generic
-18 error argv "/home/daniel/.nvm/versions/node/v6.10.1/bin/node" "/home/daniel/.nvm/versions/node/v6.10.1/bin/npm" "run" "test"
-19 error node v6.10.1
-20 error npm  v3.10.10
-21 error code ELIFECYCLE
-22 error rooset-domain-query-reducers@0.1.0 test: `jest --runInBand`
-22 error Exit status 1
-23 error Failed at the rooset-domain-query-reducers@0.1.0 test script 'jest --runInBand'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the rooset-domain-query-reducers package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     jest --runInBand
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs rooset-domain-query-reducers
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls rooset-domain-query-reducers
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
